pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aicontracts"
dynamic = ["version"]
description = "Repo-local, fail-closed guardrails for autonomous coding/build agents."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "Piyush Vyas" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
keywords = [
    "agent", "contracts", "coding-agent", "build-agent", "guardrails", "repo-local",
    "fail-closed", "budgets", "tool-use", "filesystem", "shell", "ci",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "click>=8.0",
]

[project.optional-dependencies]
otel = ["opentelemetry-api>=1.20"]
langchain = ["langchain-core>=0.2"]
crewai = ["crewai>=0.50"]
pydantic-ai = ["pydantic-ai>=0.1"]
openai = ["openai-agents==0.8.4"]
claude = ["claude-agent-sdk==0.1.50; python_version>='3.10'"]
all = [
    "aicontracts[otel]",
    "aicontracts[langchain]",
    "aicontracts[crewai]",
    "aicontracts[pydantic-ai]",
    "aicontracts[openai]",
    "aicontracts[claude]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
aicontracts = "agent_contracts.cli:main"

[project.urls]
Homepage = "https://github.com/pyyush/agentcontracts"
Documentation = "https://github.com/pyyush/agentcontracts/tree/main/spec"
Issues = "https://github.com/pyyush/agentcontracts/issues"

[tool.hatch.version]
path = "src/agent_contracts/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_contracts"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "schemas/",
    "LICENSE",
    "README.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]
ignore = ["E501"]  # Line length handled by formatter

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "yaml",
    "jsonschema",
    "jsonschema.*",
    "opentelemetry",
    "opentelemetry.*",
    "openai_agents",
    "langchain_core.*",
]
ignore_missing_imports = true
